+++ addons/ui_react/scripts/controls/ui_react_progress_bar.gd
@@ -34,6 +34,8 @@ func _ready() -> void:
 
 
 func _disconnect_all_states() -> void:
+	if _value_state != null:
+		UiReactComputedService.hook_unbind(_value_state, self, &"value_state")
 	if _value_state != null and _value_state.value_changed.is_connected(_on_value_state_changed):
 		_value_state.value_changed.disconnect(_on_value_state_changed)
 
@@ -43,6 +45,7 @@ func _connect_all_states() -> void:
 		_value_state.value_changed.connect(_on_value_state_changed)
 		_on_value_state_changed(_value_state.get_value(), _value_state.get_value())
 		_last_value = UiReactStateBindingHelper.coerce_float(_value_state.get_value())
+		UiReactComputedService.hook_bind(_value_state, self, &"value_state")
 	else:
 		_last_value = value
 	_was_completed = _is_completed()